Split copper-balance geometry into lattice and spatial modules
Pure code motion: geom/copper_balance.rs becomes a directory module
whose API, profile, and solve orchestration stay in mod.rs, with the
hex-lattice enumeration, containment classification, activation radii,
and rounded-hex template in lattice.rs, and the density kernel, cell
coverage sampling, and box-sum projection in spatial.rs. Unit tests
move with their subjects; the public surface is unchanged.
