+++ crates/pcb-ir/src/geom/copper_balance/lattice.rs
@@ -1,0 +1,442 @@
+//! Hexagonal void lattice: site enumeration and addressing, containment
+//! classification, per-site activation radii, and the rounded-hex template.
+
+use std::f64::consts::PI;
+
+use super::{DenseCopperBalanceProfile, SQRT_3};
+use crate::geom::path::transform_cmds;
+use crate::geom::{Affine2, BBox, ContourBuf, ContourSet, FillRule, PathCmd, Point, tol};
+
+const HEXAGON_CORNER_RADIUS_RATIO: f64 = 0.15;
+// A sharp regular hexagon has area 3√3 R² / 2. Rounding each 120° corner
+// inward by fillet radius kR removes (2√3 - π)k²R² across all six corners.
+pub(super) const ROUNDED_HEXAGON_AREA_FACTOR: f64 = 3.0 * SQRT_3 / 2.0
+    - (2.0 * SQRT_3 - PI) * HEXAGON_CORNER_RADIUS_RATIO * HEXAGON_CORNER_RADIUS_RATIO;
+
+#[derive(Debug, Default)]
+pub(super) struct LatticeCandidates {
+    pub(super) full_centers: Vec<Point>,
+    pub(super) edge_candidates: Vec<(Point, f64)>,
+}
+
+impl LatticeCandidates {
+    pub(super) fn new(
+        voidable: &ContourSet,
+        origin: Point,
+        profile: DenseCopperBalanceProfile,
+    ) -> Self {
+        if voidable.is_empty() {
+            return Self::default();
+        }
+
+        let candidate_region = voidable.disk_dilate(profile.max_void_radius_mm);
+        let centers = hex_aligned_lattice_centers(candidate_region.bbox, origin, profile)
+            .into_iter()
+            .filter(|center| candidate_region.contains_point(*center))
+            .collect::<Vec<_>>();
+        let fully_contained = fully_contained_hexagons(voidable, &centers, profile);
+        let mut full_centers = Vec::new();
+        let mut edge_centers = Vec::new();
+        for (center, full) in centers.into_iter().zip(fully_contained) {
+            if full {
+                full_centers.push(center);
+            } else {
+                edge_centers.push(center);
+            }
+        }
+        let edge_candidates = minimum_partial_candidates(voidable, &edge_centers, profile);
+        Self {
+            full_centers,
+            edge_candidates,
+        }
+    }
+
+    pub(super) fn is_empty(&self) -> bool {
+        self.full_centers.is_empty() && self.edge_candidates.is_empty()
+    }
+
+    fn full_void_area(&self, radius: f64) -> f64 {
+        self.full_centers.len() as f64 * ROUNDED_HEXAGON_AREA_FACTOR * radius.powi(2)
+    }
+
+    pub(super) fn partial_voids(
+        &self,
+        voidable: &ContourSet,
+        radius: f64,
+        profile: DenseCopperBalanceProfile,
+    ) -> ContourSet {
+        let candidates = self
+            .edge_candidates
+            .iter()
+            .map(|(center, activation_radius)| (*center, radius.max(*activation_radius)))
+            .collect::<Vec<_>>();
+        clipped_voids_containing_minimum_disk(voidable, &candidates, profile)
+    }
+
+    pub(super) fn void_area(
+        &self,
+        voidable: &ContourSet,
+        radius: f64,
+        profile: DenseCopperBalanceProfile,
+    ) -> f64 {
+        self.full_void_area(radius) + self.partial_voids(voidable, radius, profile).area()
+    }
+}
+
+pub(super) fn hex_aligned_lattice_centers(
+    bbox: BBox,
+    origin: Point,
+    profile: DenseCopperBalanceProfile,
+) -> Vec<Point> {
+    // Hexagon vertices are at 0°, 60°, ...; nearest-neighbor center vectors
+    // are at 30°, 90°, ... so parallel flats face each other.
+    let column_pitch = profile.lattice_column_pitch_mm();
+    let first_column = ((bbox.min.x - origin.x) / column_pitch).floor() as i64;
+    let last_column = ((bbox.max.x - origin.x) / column_pitch).ceil() as i64;
+    let mut centers = Vec::new();
+
+    for column in first_column..=last_column {
+        let x = origin.x + column as f64 * column_pitch;
+        let column_offset = if column.rem_euclid(2) == 0 {
+            0.0
+        } else {
+            profile.pitch_mm / 2.0
+        };
+        let column_origin_y = origin.y + column_offset;
+        let first_row = ((bbox.min.y - column_origin_y) / profile.pitch_mm).floor() as i64;
+        let last_row = ((bbox.max.y - column_origin_y) / profile.pitch_mm).ceil() as i64;
+
+        for row in first_row..=last_row {
+            centers.push(Point::new(
+                x,
+                column_origin_y + row as f64 * profile.pitch_mm,
+            ));
+        }
+    }
+
+    centers
+}
+
+fn minimum_partial_candidates(
+    voidable: &ContourSet,
+    centers: &[Point],
+    profile: DenseCopperBalanceProfile,
+) -> Vec<(Point, f64)> {
+    if centers.is_empty() {
+        return Vec::new();
+    }
+
+    let min_radius = profile.min_void_radius_mm;
+    let max_radius = profile.max_void_radius_mm;
+    let min_trials = uniform_candidates(centers, min_radius);
+    let max_trials = uniform_candidates(centers, max_radius);
+    let accepted_at_min = accepted_candidate_mask(voidable, &min_trials, profile);
+    let accepted_at_max = accepted_candidate_mask(voidable, &max_trials, profile);
+    let mut bounds = accepted_at_min
+        .into_iter()
+        .zip(accepted_at_max)
+        .map(|(at_min, at_max)| match (at_min, at_max) {
+            (true, _) => Some((min_radius, min_radius)),
+            (false, true) => Some((min_radius, max_radius)),
+            (false, false) => None,
+        })
+        .collect::<Vec<_>>();
+
+    loop {
+        let trials = centers
+            .iter()
+            .copied()
+            .zip(&bounds)
+            .enumerate()
+            .filter_map(|(index, (center, bounds))| {
+                let &(low, high) = bounds.as_ref()?;
+                (high - low > tol::FLATTEN_MM).then_some((index, center, (low + high) / 2.0))
+            })
+            .collect::<Vec<_>>();
+        if trials.is_empty() {
+            break;
+        }
+        let trial_geometry = trials
+            .iter()
+            .map(|(_, center, radius)| (*center, *radius))
+            .collect::<Vec<_>>();
+        let accepted = accepted_candidate_mask(voidable, &trial_geometry, profile);
+        for ((index, _, radius), accepted) in trials.into_iter().zip(accepted) {
+            let (low, high) = bounds[index].as_mut().expect("trial has radius bounds");
+            if accepted {
+                *high = radius;
+            } else {
+                *low = radius;
+            }
+        }
+    }
+
+    centers
+        .iter()
+        .copied()
+        .zip(bounds)
+        .filter_map(|(center, bounds)| bounds.map(|(_, high)| (center, high)))
+        .collect()
+}
+
+fn fully_contained_hexagons(
+    voidable: &ContourSet,
+    centers: &[Point],
+    profile: DenseCopperBalanceProfile,
+) -> Vec<bool> {
+    let candidates = uniform_candidates(centers, profile.max_void_radius_mm);
+    let outside = hexagon_set_with_radii(&candidates, voidable.tolerance).difference(voidable);
+    let outside_points = representative_points(&outside);
+    candidate_point_mask(&candidates, &outside_points, profile)
+        .into_iter()
+        .map(|has_outside_point| !has_outside_point)
+        .collect()
+}
+
+fn uniform_candidates(centers: &[Point], radius: f64) -> Vec<(Point, f64)> {
+    centers.iter().map(|center| (*center, radius)).collect()
+}
+
+fn accepted_candidate_mask(
+    voidable: &ContourSet,
+    candidates: &[(Point, f64)],
+    profile: DenseCopperBalanceProfile,
+) -> Vec<bool> {
+    let raw = hexagon_set_with_radii(candidates, voidable.tolerance).intersection(voidable);
+    let core_points = minimum_disk_core_points(&raw, voidable, candidates, profile);
+    candidate_point_mask(candidates, &core_points, profile)
+}
+
+fn clipped_voids_containing_minimum_disk(
+    voidable: &ContourSet,
+    candidates: &[(Point, f64)],
+    profile: DenseCopperBalanceProfile,
+) -> ContourSet {
+    let raw = hexagon_set_with_radii(candidates, voidable.tolerance).intersection(voidable);
+    let mut core_points = minimum_disk_core_points(&raw, voidable, candidates, profile);
+    core_points.sort_by(|left, right| left.x.total_cmp(&right.x));
+    let rings = raw
+        .connected_components()
+        .into_iter()
+        .filter(|component| component_contains_any_point(component, &core_points))
+        .flat_map(|component| component.rings)
+        .collect();
+    ContourSet::new(rings, FillRule::NonZero, voidable.tolerance)
+}
+
+fn minimum_disk_core_points(
+    raw: &ContourSet,
+    voidable: &ContourSet,
+    candidates: &[(Point, f64)],
+    profile: DenseCopperBalanceProfile,
+) -> Vec<Point> {
+    let minimum_radius = profile.minimum_partial_void_inradius_mm();
+    let mut points = representative_points(&raw.disk_erode(minimum_radius));
+    // Preserve the exact equality case, where erosion can collapse a valid
+    // minimum disk to a zero-area point that regularization omits.
+    points.extend(candidates.iter().filter_map(|(center, _)| {
+        voidable
+            .contains_disk(*center, minimum_radius)
+            .then_some(*center)
+    }));
+    points
+}
+
+fn representative_points(region: &ContourSet) -> Vec<Point> {
+    region
+        .connected_components()
+        .into_iter()
+        .filter_map(|component| component.rings.first()?.first().copied())
+        .map(|[x, y]| Point::new(x, y))
+        .collect()
+}
+
+fn candidate_point_mask(
+    candidates: &[(Point, f64)],
+    points: &[Point],
+    profile: DenseCopperBalanceProfile,
+) -> Vec<bool> {
+    let mut matched = vec![false; candidates.len()];
+    let mut by_x = candidates
+        .iter()
+        .enumerate()
+        .map(|(index, (center, radius))| (index, *center, *radius))
+        .collect::<Vec<_>>();
+    by_x.sort_by(|left, right| left.1.x.total_cmp(&right.1.x));
+    let search_radius = profile.max_void_radius_mm + tol::FLATTEN_MM;
+    for point in points {
+        let first = by_x.partition_point(|(_, center, _)| center.x < point.x - search_radius);
+        let nearest = by_x[first..]
+            .iter()
+            .take_while(|(_, center, _)| center.x <= point.x + search_radius)
+            .filter_map(|&(index, center, radius)| {
+                let distance = center.distance_to(*point);
+                (distance <= radius + tol::FLATTEN_MM).then_some((index, distance))
+            })
+            .min_by(|left, right| left.1.total_cmp(&right.1));
+        if let Some((index, _)) = nearest {
+            matched[index] = true;
+        }
+    }
+    matched
+}
+
+fn component_contains_any_point(component: &ContourSet, points: &[Point]) -> bool {
+    let first = points.partition_point(|point| point.x < component.bbox.min.x);
+    points[first..]
+        .iter()
+        .take_while(|point| point.x <= component.bbox.max.x)
+        .any(|point| {
+            point.y >= component.bbox.min.y
+                && point.y <= component.bbox.max.y
+                && component.contains_point(*point)
+        })
+}
+
+pub(super) fn hexagon_set_with_radii(candidates: &[(Point, f64)], tolerance: f64) -> ContourSet {
+    let contours = candidates
+        .iter()
+        .map(|(center, radius)| {
+            let hexagon = rounded_hexagonal_void(*radius).expect("candidate radius is validated");
+            transform_cmds(hexagon.cmds.iter().copied(), Affine2::translation(*center))
+        })
+        .collect::<Vec<_>>();
+    ContourSet::from_filled_contours(&contours, tolerance)
+}
+
+/// One slightly rounded, flat-top regular hexagonal void centered at zero.
+pub fn rounded_hexagonal_void(radius: f64) -> Option<ContourBuf> {
+    if !radius.is_finite() || radius <= 0.0 {
+        return None;
+    }
+
+    let corner_radius = radius * HEXAGON_CORNER_RADIUS_RATIO;
+    let tangent_distance = corner_radius / SQRT_3;
+    let center_inset = 2.0 * corner_radius / SQRT_3;
+    let vertices = (0..6)
+        .map(|index| {
+            let angle = index as f64 * PI / 3.0;
+            Point::new(radius * angle.cos(), radius * angle.sin())
+        })
+        .collect::<Vec<_>>();
+
+    let corner = |index: usize| {
+        let previous = vertices[(index + 5) % 6];
+        let vertex = vertices[index];
+        let next = vertices[(index + 1) % 6];
+        let incoming = vertex + (previous - vertex) * (tangent_distance / radius);
+        let outgoing = vertex + (next - vertex) * (tangent_distance / radius);
+        let center = vertex * ((radius - center_inset) / radius);
+        (incoming, outgoing, center)
+    };
+
+    let (first_incoming, _, _) = corner(0);
+    let mut commands = Vec::with_capacity(14);
+    commands.push(PathCmd::move_to(first_incoming));
+    for index in 0..6 {
+        let (incoming, outgoing, center) = corner(index);
+        if index > 0 {
+            commands.push(PathCmd::line_to(incoming));
+        }
+        commands.push(PathCmd::arc_to(outgoing, center, false));
+    }
+    commands.push(PathCmd::close());
+    Some(ContourBuf::new(commands))
+}
+
+pub(super) fn lattice_index(
+    point: Point,
+    origin: Point,
+    profile: DenseCopperBalanceProfile,
+) -> (i64, i64) {
+    let column = ((point.x - origin.x) / profile.lattice_column_pitch_mm()).round() as i64;
+    let column_offset = column.rem_euclid(2) as f64 * profile.pitch_mm / 2.0;
+    let row = ((point.y - origin.y - column_offset) / profile.pitch_mm).round() as i64;
+    (column, row)
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::geom::{BBox, ContourSet, PathOp, Point, tol};
+
+    #[test]
+    fn rejects_partial_voids_that_cannot_hold_the_minimum_disk() {
+        let profile = DenseCopperBalanceProfile::V1;
+        let voidable = ContourSet::rectangle(
+            BBox::new(Point::new(0.0, 0.0), Point::new(0.15, 10.0)),
+            tol::REGION_MM,
+        );
+        let lattice = LatticeCandidates::new(&voidable, Point::new(0.0, 0.0), profile);
+
+        assert!(lattice.edge_candidates.is_empty());
+    }
+
+    #[test]
+    fn exact_hex_containment_is_not_circumcircle_conservative() {
+        let profile = DenseCopperBalanceProfile::V1;
+        let voidable = ContourSet::rectangle(
+            BBox::new(Point::new(-0.66, -0.57), Point::new(0.66, 0.57)),
+            tol::REGION_MM,
+        );
+        let center = Point::new(0.0, 0.0);
+
+        assert!(!voidable.contains_disk(center, profile.max_void_radius_mm));
+        assert_eq!(
+            fully_contained_hexagons(&voidable, &[center], profile),
+            vec![true]
+        );
+    }
+
+    #[test]
+    fn partial_void_activation_is_monotone_in_radius() {
+        let profile = DenseCopperBalanceProfile::V1;
+        let voidable = ContourSet::rectangle(
+            BBox::new(Point::new(0.0, 0.0), Point::new(4.0, 4.0)),
+            tol::REGION_MM,
+        );
+        let centers = hex_aligned_lattice_centers(
+            voidable.disk_dilate(profile.max_void_radius_mm).bbox,
+            Point::new(0.0, 0.0),
+            profile,
+        );
+        let mut previously_accepted = vec![false; centers.len()];
+        for radius in [0.20, 0.30, 0.40, 0.50, 0.60, 0.65] {
+            let candidates = uniform_candidates(&centers, radius);
+            let accepted = accepted_candidate_mask(&voidable, &candidates, profile);
+            assert!(
+                previously_accepted
+                    .iter()
+                    .zip(&accepted)
+                    .all(|(previous, current)| !previous || *current)
+            );
+            previously_accepted = accepted;
+        }
+    }
+
+    #[test]
+    fn rounded_hexagon_uses_six_scaled_corner_arcs_and_tracks_analytic_area() {
+        let radius = 0.8;
+        let hexagon = rounded_hexagonal_void(radius).unwrap();
+        let arcs = hexagon
+            .cmds
+            .iter()
+            .filter(|command| command.op == PathOp::ArcTo)
+            .collect::<Vec<_>>();
+        assert_eq!(arcs.len(), 6);
+        for arc in arcs {
+            assert!(
+                (arc.p0.distance_to(arc.p1) - radius * HEXAGON_CORNER_RADIUS_RATIO).abs() <= 1e-12
+            );
+        }
+
+        let region = ContourSet::from_filled_contours(&[hexagon], tol::REGION_MM);
+        let expected_area = ROUNDED_HEXAGON_AREA_FACTOR * radius.powi(2);
+        assert!(
+            (region.area() - expected_area).abs() <= expected_area * 2e-3,
+            "geometric area {}, analytic area {}",
+            region.area(),
+            expected_area
+        );
+    }
+}
